requirements.txt
Django==1.10.5
django-hitcount==1.2.2
pyspark==2.1.1+hadoop2.7
